Removed paypal link
Recently, Paypal said Nintendo is being much more agressive, not allowing even donations to people who make parody games with some of their content.
Even if we developers make entirely new content from scratch, if it resembles some of Nintendo characters, now it refuses to allow indie gamers to receive donations!!

That is, we have a give person  that loves a given theme (in a way some characters are WAIFUs), then he/she uses most of his/her idle times (sometimes entire nights) to make a game like this, and then we are refused to receive anything, even a $1 thank you, from people who got touched by the game somehow.

Its a shame, but rules are rules. As we are all game lovers (and as I started this project with the word freeware in mind), be it unicque games or original games, we want to fantasize (our minds are law free at the moment). So I hope this will just increase the parody content, in our free times, to show 'them' that we don't need money and we don't want to 'compete' with great corporations. We all payed for 'their' content at least once (nintendo cartridges, sega cds, and else) anyway.

We just want to play games the way we wanted (more freedom, more characters, etc).

Thats it.